feat(utils): move parts from utils into core
