+++ packages/cli/src/lib/collect/command-object.spec.ts
@@ -1,6 +1,6 @@
 import { Report } from '@quality-metrics/models';
 import { dummyConfig } from '@quality-metrics/models/testing';
-import { CollectOptions } from '@quality-metrics/utils';
+import { CollectOptions } from '@quality-metrics/core';
 import { readFileSync } from 'node:fs';
 import { join } from 'node:path';
 import { yargsCli } from '../cli';
