+++ packages/cli/src/lib/collect/command-object.ts
@@ -2,10 +2,9 @@ import { pluginOutputSchema, Report } from '@quality-metrics/models';
 import {
   collect,
   CollectOptions,
-  CollectOutputError,
   persistReport,
   logPersistedResults,
-} from '@quality-metrics/utils';
+} from '@quality-metrics/core';
 import { CommandModule } from 'yargs';
 import * as packageJson from '../../../package.json';
 
@@ -30,7 +29,7 @@ export function yargsCollectCommandObject() {
         // Running checks after persisting helps while debugging as you can check the invalid output after the error
         pluginOutputSchema.parse(plugin);
       } catch (e) {
-        throw new CollectOutputError(plugin.slug, e as Error);
+        throw new Error(plugin.slug + (e as Error).message);
       }
     });
   };
