+++ packages/core/src/index.ts
@@ -1,1 +1,3 @@
-export * from './lib/core';
+export { logPersistedResults, persistReport } from './lib/persist';
+export { executePlugins } from './lib/execute-plugin';
+export { collect, CollectOptions } from './lib/collect';
