--- packages/core/src/lib/core.spec.ts
@@ -1,7 +1,0 @@
-import { core } from './core';
-
-describe('core', () => {
-  it('should work', () => {
-    expect(core()).toEqual('core');
-  });
-});
