--- packages/core/src/lib/core.ts
@@ -1,3 +1,0 @@
-export function core(): string {
-  return 'core';
-}
