+++ packages/utils/src/index.ts
@@ -6,14 +6,6 @@ export {
   ProcessError,
   objectToCliArgs,
 } from './lib/collect/implementation/execute-process';
-export {
-  collect,
-  CollectOptions,
-  CollectOutputError,
-} from './lib/collect/index';
-export {
-  persistReport,
-  PersistDirError,
-  PersistError,
-  logPersistedResults,
-} from './lib/collect/implementation/persist';
+export { calcDuration, formatBytes } from './lib/collect/implementation/utils';
+export { reportToStdout } from './lib/collect/implementation/report-to-stdout';
+export { reportToMd } from './lib/collect/implementation/report-to-md';
