+++ tsconfig.base.json
@@ -18,15 +18,15 @@
     "baseUrl": ".",
     "resolveJsonModule": true,
     "paths": {
-      "@quality-metrics-cli/core": ["packages/core/src/index.ts"],
-      "@quality-metrics-cli/nx-plugin": ["packages/nx-plugin/src/index.ts"],
       "@quality-metrics/cli": ["packages/cli/src/index.ts"],
+      "@quality-metrics/core": ["packages/core/src/index.ts"],
       "@quality-metrics/eslint-plugin": ["packages/plugin-eslint/src/index.ts"],
       "@quality-metrics/lighthouse-plugin": [
         "packages/plugin-lighthouse/src/index.ts"
       ],
       "@quality-metrics/models": ["packages/models/src/index.ts"],
       "@quality-metrics/models/testing": ["packages/models/test/index.ts"],
+      "@quality-metrics/nx-plugin": ["packages/nx-plugin/src/index.ts"],
       "@quality-metrics/utils": ["packages/utils/src/index.ts"]
     }
   },
